+++ ProgC/Lab_12_1_2/mylib.h
@@ -11,7 +11,7 @@
 
 int isSquare(long long int x);
 
-ARR_DLL int ARR_DECL get_new_mas(double *src, int src_len, double *dst, int *dst_len);
+ARR_DLL int ARR_DECL get_new_mas(int *src, int src_len, int *dst, int *dst_len);
 
 ARR_DLL void ARR_DECL sdvig(int* mas, int n, int k);
